+++ designpattern/src/com/org/decorator/impl/addons/Whip.java
@@ -1,19 +1,19 @@
-package com.org.decorator.impl;
+package com.org.decorator.impl.addons;
 
 import com.org.decorator.Beverage;
 import com.org.decorator.CondimentDecorator;
 
-public class Mocha extends CondimentDecorator {
+public class Whip extends CondimentDecorator {
 
-    private final double COST = 0.20;
-    public Beverage beverage;
+    private final double COST = 0.75;
+    Beverage beverage;
 
-    public Mocha(Beverage beverage){
+    public Whip (Beverage beverage){
         this.beverage = beverage;
     }
     @Override
     public String getDescription() {
-        return beverage.getDescription() + ", Mocha";
+        return beverage.getDescription() + ", Whip";
     }
 
     public double cost (){
